ports: support resolver tunnel factories on the web (#186937)
@alexr00 I just copied the non-Node.js-specific code into the base port
class and things seem to work. What do you think?

Fixes https://github.com/microsoft/vscode-remote-release/issues/8595
